requirements.txt
agency-swarm==0.3.1
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==1.10.3
beautifulsoup4==4.8.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datamodel-code-generator==0.26.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fsspec==2024.10.0
genson==1.3.0
gradio==5.5.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonref==1.1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.3
openai==1.51.2
ordered-set==4.1.0
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pinecone==5.4.0
pinecone-plugin-inference==2.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.7.1
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
sympy==1.13.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.1.0
tomlkit==0.12.0
tqdm==4.67.0
transformers==4.46.2
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
